Substantial updates to Text-Adventure.py and associated txt files
Consolidated the code for reading commands.txt and commands_text.txt into one function. Updated general txt reading function to evaluate inputs, to allow for integer, boolean, etc. dictionary entries. Added an object system, allowing objects to be defined and allocated to a room, and a 'look' system to allow the player to see a more detailed description of an object. Merged the "connect_direction" and "door_direction" entries in the room txt files as only one is actually needed. Renamed room and object txt files.